requirements.txt.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.2
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
Flask==3.0.3
Flask-Mail==0.10.0
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.23.4
idna==3.7
image==1.5.33
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyngrok==7.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.1.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
starlette==0.37.2
sympy==1.13.0
tbb==2021.13.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.5
Werkzeug==3.0.3
